Rename ErrorDetailCollection --> ErrorMessage::DetailCollection
More accurately reflects its intention and groups
it with its co-dependendants.
