+++ .rubocop_todo.yml
@@ -305,7 +305,7 @@ Metrics/MethodLength:
     - 'app/models/timed_transitions/transitioner.rb'
     - 'app/presenters/claim_state_transition_presenter.rb'
     - 'app/presenters/error_message_translator.rb'
-    - 'app/presenters/error_presenter.rb'
+    - 'app/presenters/error_message/presenter.rb'
     - 'app/presenters/expense_type_presenter.rb'
     - 'app/presenters/fee/interim_fee_type_presenter.rb'
     - 'app/rake_helpers/cloned_claim_repairer.rb'
@@ -448,7 +448,7 @@ Naming/PredicateName:
     - 'app/models/supplier_number.rb'
     - 'app/models/timed_transitions/transitioner.rb'
     - 'app/presenters/claim/base_claim_presenter.rb'
-    - 'app/presenters/error_presenter.rb'
+    - 'app/presenters/error_message/presenter.rb'
     - 'app/presenters/fee/interim_fee_type_presenter.rb'
     - 'app/services/remote/simple_associations.rb'
     - 'app/services/slack_notifier.rb'
@@ -678,7 +678,7 @@ RSpec/ContextMethod:
     - 'spec/models/fee/base_fee_type_spec.rb'
     - 'spec/models/google_analytics/data_tracking_spec.rb'
     - 'spec/models/provider_spec.rb'
-    - 'spec/presenters/error_presenter_spec.rb'
+    - 'spec/presenters/error_message/presenter_spec.rb'
     - 'spec/presenters/fixed_fee_presenter_spec.rb'
     - 'spec/presenters/interim_fee_presenter_spec.rb'
     - 'spec/presenters/management_information_presenter_spec.rb'
@@ -823,7 +823,7 @@ RSpec/ContextWording:
     - 'spec/presenters/claim_state_transition_presenter_spec.rb'
     - 'spec/presenters/error_detail_collection_spec.rb'
     - 'spec/presenters/error_message_translator_spec.rb'
-    - 'spec/presenters/error_presenter_spec.rb'
+    - 'spec/presenters/error_message/presenter_spec.rb'
     - 'spec/presenters/expense_presenter_spec.rb'
     - 'spec/presenters/expense_type_presenter_spec.rb'
     - 'spec/presenters/fee/basic_fee_presenter_spec.rb'
@@ -975,7 +975,7 @@ RSpec/DescribedClass:
     - 'spec/presenters/error_detail_collection_spec.rb'
     - 'spec/presenters/error_detail_spec.rb'
     - 'spec/presenters/error_message_translator_spec.rb'
-    - 'spec/presenters/error_presenter_spec.rb'
+    - 'spec/presenters/error_message/presenter_spec.rb'
     - 'spec/presenters/expense_presenter_spec.rb'
     - 'spec/presenters/expense_type_presenter_spec.rb'
     - 'spec/presenters/fee_presenter_spec.rb'
@@ -1326,7 +1326,7 @@ RSpec/ExampleLength:
     - 'spec/presenters/claim/transfer_claim_presenter_spec.rb'
     - 'spec/presenters/determination_presenter_spec.rb'
     - 'spec/presenters/error_message_translator_spec.rb'
-    - 'spec/presenters/error_presenter_spec.rb'
+    - 'spec/presenters/error_message/presenter_spec.rb'
     - 'spec/presenters/redetermination_presenter_spec.rb'
     - 'spec/services/claims/case_worker_claim_updater_spec.rb'
     - 'spec/services/claims/create_draft_spec.rb'
@@ -2177,7 +2177,7 @@ RSpec/NestedGroups:
     - 'spec/models/representation_order_spec.rb'
     - 'spec/models/stage_collection_spec.rb'
     - 'spec/models/timed_transitions/transitioner_spec.rb'
-    - 'spec/presenters/error_presenter_spec.rb'
+    - 'spec/presenters/error_message/presenter_spec.rb'
     - 'spec/presenters/expense_presenter_spec.rb'
     - 'spec/presenters/fee/basic_fee_presenter_spec.rb'
     - 'spec/services/cclf/case_type_adapter_spec.rb'
@@ -2586,7 +2586,7 @@ Rails/FilePath:
     - 'app/models/claim_json_schema_validator.rb'
     - 'app/models/claim_state_transition_reason.rb'
     - 'app/models/stats/collector/time_to_completion_collector.rb'
-    - 'app/presenters/error_presenter.rb'
+    - 'app/presenters/error_messages/presenter.rb'
     - 'app/rake_helpers/clone_repair_runner.rb'
     - 'lib/data_migrator/provider_suppliers_migrator.rb'
     - 'lib/extensions/rails_module_extension.rb'
@@ -2594,6 +2594,7 @@ Rails/FilePath:
     - 'spec/lib/working_day_calculator_spec.rb'
     - 'spec/models/claim/advocate_claim_spec.rb'
     - 'spec/models/stats/collector/time_to_completion_collector_spec.rb'
+    - 'spec/presenters/error_message/presenter_spec.rb'
     - 'spec/rails_helper.rb'
     - 'spec/support/api/api_test_client.rb'
     - 'spec/support/matchers/validation_matchers.rb'
@@ -3147,7 +3148,7 @@ Style/CommentAnnotation:
   Exclude:
     - 'app/models/claims/sort.rb'
     - 'app/models/fee/base_fee.rb'
-    - 'app/presenters/error_presenter.rb'
+    - 'app/presenters/error_message/presenter.rb'
     - 'spec/i18n_spec.rb'
     - 'spec/models/claims/financial_summary_spec.rb'
     - 'spec/services/claims/case_worker_claim_updater_spec.rb'
@@ -3399,7 +3400,7 @@ Style/NumericLiterals:
     - 'spec/models/fee/base_fee_spec.rb'
     - 'spec/models/timed_transitions/batch_transitioner_spec.rb'
     - 'spec/presenters/error_detail_spec.rb'
-    - 'spec/presenters/error_presenter_spec.rb'
+    - 'spec/presenters/error_message/presenter_spec.rb'
     - 'spec/presenters/expense_presenter_spec.rb'
     - 'spec/presenters/fee_presenter_spec.rb'
     - 'spec/services/claims/external_user_claim_updater_spec.rb'
@@ -3829,7 +3830,7 @@ Layout/LineLength:
     - 'spec/presenters/disbursement_presenter_spec.rb'
     - 'spec/presenters/error_detail_collection_spec.rb'
     - 'spec/presenters/error_message_translator_spec.rb'
-    - 'spec/presenters/error_presenter_spec.rb'
+    - 'spec/presenters/error_message/presenter_spec.rb'
     - 'spec/presenters/expense_presenter_spec.rb'
     - 'spec/presenters/expense_type_presenter_spec.rb'
     - 'spec/presenters/fee_presenter_spec.rb'
